Hides Neo4J Graph if Custom Query was used
Hides Neo4J Graph if Custom Query was used.
Utilizes Local Storage Token, Basically stops Neo4J from querying backend if storage token is set to be using custom query.

The Custom query value is changed when a project is loaded normally to TRUE.
The Custom query value is changed when a custom query project is loaded to FALSE.